multi-cluster implementation of Dispatcher.selectForDispatch (#150)
Adapt code from multicluster branch to current selectForDispatch
structure to enable cluster selection and multi-cluster dispatching.
As in the multicluster branch, it is assumed that an external
mechanism is used to assign AppWrappers to target clusters. The
dispatching logic simply loops over all known clusters and filters
AppWrappers based on their assigned cluster.